+++ libft/ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: scoron <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/13 17:52:30 by scoron            #+#    #+#             */
-/*   Updated: 2019/01/14 18:21:56 by scoron           ###   ########.fr       */
+/*   Updated: 2019/01/21 18:53:43 by scoron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static char	*ft_putint(char *res, long long n, size_t len)
+static char	*ft_putint(char *res, intmax_t n, size_t len)
 {
 	if (n == 0)
 		res[0] = '0';
@@ -34,11 +34,11 @@ static char	*ft_putint(char *res, long long n, size_t len)
 	return (res);
 }
 
-char		*ft_itoa(long long n)
+char		*ft_itoa(intmax_t n)
 {
 	char		*res;
 	size_t		len;
-	long long	k;
+	intmax_t	k;
 
 	len = 1;
 	k = n;
